+++ src/com/kaylerrenslow/armaDialogCreator/gui/fxcontrol/treeView/TreeItemData.java
@@ -1,6 +1,8 @@
 package com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.treeView;
 
 
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.scene.Node;
@@ -9,10 +11,6 @@ import org.jetbrains.annotations.Nullable;
 
 public class TreeItemData {
 
-	private static int lastId = 0;
-
-	private final int ID = lastId++;
-
 	private final CellType cellType;
 	private TreeNodeUpdateListener updateListener;
 
@@ -20,12 +18,14 @@ public class TreeItemData {
 
 	private String text;
 	private final ObservableList<String> styleClassList = FXCollections.observableArrayList();
+	private SimpleStringProperty textProperty = new SimpleStringProperty();
 
 
 	public TreeItemData(@NotNull String text, @NotNull CellType cellType, @Nullable Node graphic) {
 		this.graphic = graphic;
 		this.text = text;
 		this.cellType = cellType;
+		textProperty.set(text);
 	}
 
 	public void setUpdateListener(TreeNodeUpdateListener updateListener) {
@@ -56,23 +56,17 @@ public class TreeItemData {
 		return cellType == CellType.COMPOSITE;
 	}
 
-
-	@Override
-	public boolean equals(Object other) {
-		return other instanceof TreeItemData && this.ID == ((TreeItemData) other).ID;
-	}
-
 	@Override
 	public final String toString() {
-		return text;
+		return textProperty.get();
 	}
 
 	public final String getText() {
-		return text;
+		return textProperty.get();
 	}
 
-	void setText(String t) {
-		this.text = t;
+	public void setText(@NotNull String t) {
+		textProperty.set(t);
 		if (this.updateListener != null) {
 			this.updateListener.renamed(t);
 		}
@@ -87,5 +81,10 @@ public class TreeItemData {
 	@NotNull
 	public ObservableList<String> getStyleClass() {
 		return styleClassList;
+	}
+
+	@NotNull
+	public ObservableValue<String> getTextObservable() {
+		return textProperty;
 	}
 }